backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.116.1
uvicorn[standard]==0.35.0

# Database
motor==3.7.1
pymongo==4.13.2

# Authentication and Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Data validation and serialization
pydantic[email]==2.11.7
email-validator==2.2.0

# Environment variables
python-dotenv==1.1.1

# HTTP client (for testing and external APIs)
requests==2.31.0

# Optional: For future AI integration
# openai==1.3.0

# Optional: For data analysis and ML
# scikit-learn==1.3.0
# pandas==2.1.0
# numpy==1.24.0

# Development and testing
pytest==7.4.0
pytest-asyncio==0.21.0
httpx==0.25.0
